Primed program for gradient calculations
Since I'm going to be using the open weather api to get the times of sunrise and sunset I created a function which is passed down to the weather module so the information from the api call is accessable from a higher level. An alternative to this is to add the sunrise/set times as properties of the weather module which can be accessed dynamically.

Potential future problem: changing the background color of the container will require refreshing the entire page, i.e. all the elements. This would likely break some interactive aespects of the elements like the news elements scroll animation and so a refresh should probably be avoided. If the setstate command doesn't interrupt these things though, we're in business!